requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
flask>=3.0.0
flask-cors>=4.0.0

# ── Authentication ────────────────────────────────────────────────────────────
flask-jwt-extended>=4.6.0
bcrypt>=4.1.0

# ── Database ──────────────────────────────────────────────────────────────────
pymongo>=4.6.0
dnspython>=2.4.0          # Required for MongoDB Atlas SRV connection strings

# ── Rate Limiting ─────────────────────────────────────────────────────────────
flask-limiter>=3.5.0

# ── ML Inference ─────────────────────────────────────────────────────────────
onnxruntime>=1.17.0       # CPU inference for ONNX FP16 model
numpy>=1.26.0
Pillow>=10.2.0            # Image preprocessing

# ── API Documentation ─────────────────────────────────────────────────────────
flasgger>=0.9.7.1         # Swagger UI at /apidocs

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0      # .env file support
